requirements.txt
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
torchmetrics
torchmetrics[audio]
numpy==1.26.4
inflect
tqdm
matplotlib
pandas
wandb
comet_ml
hydra-core==1.3.2

black
isort
pre-commit